requirements.txt
Scrapy==2.0.0
scrapyd-client==1.1
python-dateutil==2.8.1
textract==1.6.3
elasticsearch-dsl==7.1
bump2version==1.0.1
elasticsearch==7.1
enum34==1.1.3
pycparser==2.14
queuelib==1.4.2
six==1.12.0
Twisted==19.10.0
urllib3==1.24.3
w3lib==1.21
PyPDF2==1.26.0
tika==1.24
firebase-admin==4.2.0
python-scrapyd-api
scrapy-rotating-proxies
schedule
datefinder==0.7.1
scrapy-splash==0.7.2
dateparser==1.0.0
textblob==0.15.3
pycountry==20.7.3
python-dotenv==0.9.1
bs4==0.0.1
langid==1.1.6
selenium==3.141.0
scrapy_selenium==0.0.7
pandas==1.1.5
openpyxl
PyMuPDF==1.16.14
pytesseract==0.3.8
fitz==0.0.1.dev2
unittest-xml-reporting==2.2.0
